+++ Assets/Scripts/SceneModel/Movement/Face/FaceResolver.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4eabe60812836b644949c96bb5f83cc5
+guid: 4903f10fbe2f22f4386e4a3730262327
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
