+++ Assets/Scripts/Settings/BoneSettings.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a58ab7b6b57c9104391f1b72e989df10
+guid: db0a9669eb6d71442a9dfd53c08c3fec
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
