+++ Assets/Scripts/Settings/Settings.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 4e53a3bf80633964688f09e8b2af2371
+guid: 80a22fae9feb70e40ab70b384d1ea62d
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
