+++ Assets/Scripts/Settings/SettingsUtil.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 73e3229810e76804dae5dd4def5ce21a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
