+++ Assets/Scripts/Settings/Types.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 908074ab04daee042bc1e429f199e5e5
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
