+++ Assets/Scripts/Settings/Types/BoolField.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5fc81232961e0cb4faeba940f8c4e22c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
