+++ Assets/Scripts/Settings/Types/EnumField.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c7a3fdad0f9defb4bb73d4cd76c7a205
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
