+++ Assets/Scripts/Settings/Types/FloatField.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 894dc2c361246204cbdfeb7411c7ac40
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
