+++ Assets/Scripts/Settings/Types/IField.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 12e397ec72f3d954a959f67c805888b9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
