+++ Assets/Scripts/Settings/Types/IntField.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e865d4b92340ab04d85b0c0fae6abe56
+guid: 3256d42b7b9fe4d4ea5931a4bfe9226a
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
