+++ Assets/Scripts/Settings/Types/TextField.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: beff49af1716fe44daa25293fd1b82d3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
